refactor(config): derive MCP server types from schema
